Split animation code from .h files into .cpps. Optimize passing variables to Bones.
