+++ src/Bone.h
@@ -1,6 +1,5 @@
 #pragma once
 #include <glm/glm.hpp>
-//#include <glm/gtc/quaternion.hpp>
 #include <glm/gtx/quaternion.hpp>
 #include <vector>
 #include <string>
@@ -24,9 +23,6 @@ struct KeyScale
     float timeStamp;
 };
 
-glm::vec3 GetGLMVec(const aiVector3D& vec);
-glm::quat GetGLMQuat1(const aiQuaternion& pOrientation);
-
 class Bone
 {
     private:
@@ -42,167 +38,46 @@ class Bone
         int m_ID;
     
     public:
+        /*reads keyframes from aiNodeAnim*/
+        Bone(const std::string& name, int ID, const aiNodeAnim* channel);
 
-/*reads keyframes from aiNodeAnim*/
-    Bone(const std::string& name, int ID, const aiNodeAnim* channel)
-        :
-        m_Name(name),
-        m_ID(ID),
-        m_LocalTransform(1.0f)
-    {
-        m_NumPositions = channel->mNumPositionKeys;
+        /*interpolates  b/w positions,rotations & scaling keys based on the curren time of 
+        the animation and prepares the local transformation matrix by combining all keys 
+        tranformations*/
+        void Update(float animationTime);
 
-        for (int positionIndex = 0; positionIndex < m_NumPositions; ++positionIndex)
-        {
-            aiVector3D aiPosition = channel->mPositionKeys[positionIndex].mValue;
-            float timeStamp = channel->mPositionKeys[positionIndex].mTime;
-            KeyPosition data;
-            data.position = GetGLMVec(aiPosition);
-            data.timeStamp = timeStamp;
-            m_Positions.push_back(data);
-        }
+        glm::mat4 GetLocalTransform() const { return m_LocalTransform; }
 
-        m_NumRotations = channel->mNumRotationKeys;
-        for (int rotationIndex = 0; rotationIndex < m_NumRotations; ++rotationIndex)
-        {
-            aiQuaternion aiOrientation = channel->mRotationKeys[rotationIndex].mValue;
-            float timeStamp = channel->mRotationKeys[rotationIndex].mTime;
-            KeyRotation data;
-            data.orientation = GetGLMQuat1(aiOrientation);
-            data.timeStamp = timeStamp;
-            m_Rotations.push_back(data);
-        }
+        std::string GetBoneName() const { return m_Name; }
 
-        m_NumScalings = channel->mNumScalingKeys;
-        for (int keyIndex = 0; keyIndex < m_NumScalings; ++keyIndex)
-        {
-            aiVector3D scale = channel->mScalingKeys[keyIndex].mValue;
-            float timeStamp = channel->mScalingKeys[keyIndex].mTime;
-            KeyScale data;
-            data.scale = GetGLMVec(scale);
-            data.timeStamp = timeStamp;
-            m_Scales.push_back(data);
-        }
-    }
-	
-    /*interpolates  b/w positions,rotations & scaling keys based on the curren time of 
-    the animation and prepares the local transformation matrix by combining all keys 
-    tranformations*/
-    void Update(float animationTime)
-    {
-        glm::mat4 translation = InterpolatePosition(animationTime);
-        glm::mat4 rotation = InterpolateRotation(animationTime);
-        glm::mat4 scale = InterpolateScaling(animationTime);
-        m_LocalTransform = translation * rotation * scale;
-    }
+        int GetBoneID() const { return m_ID; }
 
-    glm::mat4 GetLocalTransform() const { return m_LocalTransform; }
-    std::string GetBoneName() const { return m_Name; }
-    int GetBoneID() const { return m_ID; }
-	
+        /* Gets the current index on mKeyPositions to interpolate to based on 
+        the current animation time*/
+        int GetPositionIndex(float animationTime) const;
 
-    /* Gets the current index on mKeyPositions to interpolate to based on 
-    the current animation time*/
-    int GetPositionIndex(float animationTime) const
-    {
-        for (int index = 0; index < m_NumPositions - 1; ++index)
-        {
-            if (animationTime < m_Positions[index + 1].timeStamp)
-                return index;
-        }
-        assert(0);
-        return 0;
-    }
+        /* Gets the current index on mKeyRotations to interpolate to based on the 
+        current animation time*/
+        int GetRotationIndex(float animationTime) const;
 
-    /* Gets the current index on mKeyRotations to interpolate to based on the 
-    current animation time*/
-    int GetRotationIndex(float animationTime) const
-    {
-        for (int index = 0; index < m_NumRotations - 1; ++index)
-        {
-            if (animationTime < m_Rotations[index + 1].timeStamp)
-                return index;
-        }
-        assert(0);
-        return 0;
-    }
-
-    /* Gets the current index on mKeyScalings to interpolate to based on the 
-    current animation time */
-    int GetScaleIndex(float animationTime) const
-    {
-        for (int index = 0; index < m_NumScalings - 1; ++index)
-        {
-            if (animationTime < m_Scales[index + 1].timeStamp)
-                return index;
-        }
-        assert(0);
-        return 0;
-    }
+        /* Gets the current index on mKeyScalings to interpolate to based on the 
+        current animation time */
+        int GetScaleIndex(float animationTime) const;
 
 private:
-
     /* Gets normalized value for Lerp & Slerp*/
-    float GetScaleFactor(float lastTimeStamp, float nextTimeStamp, float animationTime) const
-    {
-        float scaleFactor = 0.0f;
-        float midWayLength = animationTime - lastTimeStamp;
-        float framesDiff = nextTimeStamp - lastTimeStamp;
-        scaleFactor = midWayLength / framesDiff;
-        return scaleFactor;
-    }
-
+    float GetScaleFactor(float lastTimeStamp, float nextTimeStamp, float animationTime) const;
+    
     /*figures out which position keys to interpolate b/w and performs the interpolation 
     and returns the translation matrix*/
-    glm::mat4 InterpolatePosition(float animationTime)
-    {
-        if (1 == m_NumPositions)
-            return glm::translate(glm::mat4(1.0f), m_Positions[0].position);
-
-        int p0Index = GetPositionIndex(animationTime);
-        int p1Index = p0Index + 1;
-        float scaleFactor = GetScaleFactor(m_Positions[p0Index].timeStamp,
-            m_Positions[p1Index].timeStamp, animationTime);
-        glm::vec3 finalPosition = glm::mix(m_Positions[p0Index].position,
-            m_Positions[p1Index].position, scaleFactor);
-        return glm::translate(glm::mat4(1.0f), finalPosition);
-    }
+    glm::mat4 InterpolatePosition(float animationTime);
 
     /*figures out which rotations keys to interpolate b/w and performs the interpolation 
     and returns the rotation matrix*/
-    glm::mat4 InterpolateRotation(float animationTime)
-    {
-        if (1 == m_NumRotations)
-        {
-            auto rotation = glm::normalize(m_Rotations[0].orientation);
-            return glm::toMat4(rotation);
-        }
-
-        int p0Index = GetRotationIndex(animationTime);
-        int p1Index = p0Index + 1;
-        float scaleFactor = GetScaleFactor(m_Rotations[p0Index].timeStamp,
-            m_Rotations[p1Index].timeStamp, animationTime);
-        glm::quat finalRotation = glm::slerp(m_Rotations[p0Index].orientation,
-            m_Rotations[p1Index].orientation, scaleFactor);
-        finalRotation = glm::normalize(finalRotation);
-        return glm::toMat4(finalRotation);
-    }
+    glm::mat4 InterpolateRotation(float animationTime);
 
     /*figures out which scaling keys to interpolate b/w and performs the interpolation 
     and returns the scale matrix*/
-    glm::mat4 InterpolateScaling(float animationTime)
-    {
-        if (1 == m_NumScalings)
-            return glm::scale(glm::mat4(1.0f), m_Scales[0].scale);
-
-        int p0Index = GetScaleIndex(animationTime);
-        int p1Index = p0Index + 1;
-        float scaleFactor = GetScaleFactor(m_Scales[p0Index].timeStamp,
-            m_Scales[p1Index].timeStamp, animationTime);
-        glm::vec3 finalScale = glm::mix(m_Scales[p0Index].scale, m_Scales[p1Index].scale
-            , scaleFactor);
-        return glm::scale(glm::mat4(1.0f), finalScale);
-    }
-
+    glm::mat4 InterpolateScaling(float animationTime);
 
 };
