+++ src/utility.h
@@ -2,7 +2,7 @@
 #define UTILITY_H_
 
 #include <glm/glm.hpp>
-
+#include <assimp/scene.h>
 #include <vector>
 
 // Initialize empty matrix made of std::vector 
@@ -20,5 +20,7 @@ void initializeMatrix(std::vector<std::vector<T>>& matrix, int rows, int cols, T
 // Resulting Vector is a vector starting in A vertex 
 // Assume: A, B,C are given in counter-clockwise order
 glm::vec3 calculateTriangleNormal(glm::vec3 A, glm::vec3 B, glm::vec3 C);
+glm::mat4 ConvertMatrixToGLMFormat(const aiMatrix4x4& from);
+
 
 #endif
